requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
install==1.3.5
mysql-connector-python==8.3.0
pyTelegramBotAPI==4.16.1
requests==2.31.0
urllib3==2.2.1
setuptools==65.5.0
pip==23.1.2